+++ flutter_appauth/ios/flutter_appauth/Sources/flutter_appauth/FlutterAppauthPlugin.m
@@ -484,7 +484,8 @@ AppAuthAuthorization *authorization;
             openURL:(NSURL *)url
             options:
                 (NSDictionary<UIApplicationOpenURLOptionsKey, id> *)options {
-  if ([_currentAuthorizationFlow resumeExternalUserAgentFlowWithURL:url]) {
+  if ([_currentAuthorizationFlow resumeExternalUserAgentFlowWithURL:url
+                                                              error:nil]) {
     _currentAuthorizationFlow = nil;
     return YES;
   }
@@ -503,7 +504,8 @@ AppAuthAuthorization *authorization;
     openURLContexts:(NSSet<UIOpenURLContext *> *)URLContexts {
   for (UIOpenURLContext *URLContext in URLContexts) {
     if ([_currentAuthorizationFlow
-            resumeExternalUserAgentFlowWithURL:URLContext.URL]) {
+            resumeExternalUserAgentFlowWithURL:URLContext.URL
+                                         error:nil]) {
       _currentAuthorizationFlow = nil;
       return YES;
     }