+++ flutter_appauth/ios/flutter_appauth/Sources/flutter_appauth/OIDExternalUserAgentIOSNoSSO.m
@@ -100,7 +100,8 @@ NS_ASSUME_NONNULL_BEGIN
         }
         strongSelf->_webAuthenticationVC = nil;
         if (callbackURL) {
-          [strongSelf->_session resumeExternalUserAgentFlowWithURL:callbackURL];
+          [strongSelf->_session resumeExternalUserAgentFlowWithURL:callbackURL
+                                                             error:nil];
         } else {
           NSError *safariError = [OIDErrorUtilities
                 errorWithCode:OIDErrorCodeUserCanceledAuthorizationFlow
@@ -182,7 +183,8 @@ NS_ASSUME_NONNULL_BEGIN
                 strongSelf->_authenticationVC = nil;
                 if (callbackURL) {
                   [strongSelf->_session
-                      resumeExternalUserAgentFlowWithURL:callbackURL];
+                      resumeExternalUserAgentFlowWithURL:callbackURL
+                                                   error:nil];
                 } else {
                   NSError *safariError = [OIDErrorUtilities
                         errorWithCode:OIDErrorCodeUserCanceledAuthorizationFlow